% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{concatfiles}
\alias{concatfiles}
\title{concat files in ones}
\usage{
concatfiles(pattern = NULL, fileout = NULL)
}
\arguments{
\item{pattern}{pattern to scan files to be merged}

\item{fileout}{filename to save the output}
}
\description{
Concat files in one file.
}
\examples{
concatfiles(pattern="ffff_001_01_apmt#[[:digit:]]{2}.ini",fileout="ffff_001_01_apmt.ini")

}
