% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_CheckDataCount}
\alias{cts5_CheckDataCount}
\title{Compare the number of available data point}
\usage{
cts5_CheckDataCount(technical, dataMerged)
}
\arguments{
\item{technical}{technical file read from \code{\link{cts5_readtechnical}}}

\item{dataMerged}{merged data read from \code{\link{cts5_concatProfile}}}
}
\value{
list containing check (= True if the number of data point are the same);
and DataCheck a data.frame with the full comparison.
}
\description{
cts5_CheckDataCount compare the number of data point per sensor and per phase
between the technical file and the dataMerged file
}
\examples{
cts5_decode(floatname=floatname,CycleNumber=c,PatternNumber = p,subdir="./CSV",sensors=c("sbe41","do","eco","ocr"))

tech<-cts5_readtechnical(floatname=floatname,CycleNumber=c,PatternNumber = p)

setwd("./CSV")

dataMerged<-cts5_concatProfile(floatname=floatname,CycleNumber=c,PatternNumber = p,
                              sensors=c("sbe41","do","eco","ocr"))

dataMerged<-cts5_ProcessData(Meta$SENSORS,dataMerged)

if (!cts5_CheckDataCount(tech,dataMerged)$check){
 cat("!! Warning, data count error \n")
}

PlotCTS5(login=login,dataMerged,PhaseToPlot=c("PRE","DES","PAR","ASC","SUR"),add=FALSE,technical=TRUE,paper = "A4",mfrow=c(3,2))

}
