% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_concatProfile}
\alias{cts5_concatProfile}
\title{cts5_concatProfile : Concat .csv files to one data.frame}
\usage{
cts5_concatProfile(
  floatname = "ffff",
  CycleNumber,
  PatternNumber = 1,
  sensors = CTS5_supported_sensors,
  dec = "."
)
}
\arguments{
\item{floatname}{hexa name of the float}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensors}{list of sensor to decode}

\item{dec}{decimal character in ASCII}
}
\value{
data.frame containing the data
}
\description{
read NKE .csv ASCII files obtained from \code{\link{cts5_decode}} and 
concat them into one data.frame
}
\details{
this function add an identification for the sensorType and acquisition phase 
as explained in \code{\link{cts5_readcsv}}
}
\examples{
dataMerged<-usea_concatProfile(floatname="ffff",CycleNumber=275)


}
