% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_convert2Ascii}
\alias{cts5_convert2Ascii}
\title{convert dataprofile list object to one data.frame}
\usage{
cts5_convert2Ascii(dataprofile, AddGPS = T)
}
\arguments{
\item{dataprofile}{data list from \code{\link{cts5_readprofile}} or \code{\link{cts5_ProcessData}}}

\item{AddGPS}{if TRUE, add GPS from technical}
}
\value{
data.frame containing the data
}
\description{
convert dataprofile list object to one data.frame
}
\details{
this function add an identification for the sensorType and acquisition phase 
as explained in \code{\link{cts5_readcsv}}
}
\examples{

login="lovuse001a"

Meta<-cts5_readMetaSensor()

cts5_decode(CycleNumber=c,PatternNumber = p,subdir="./CSV")

dataprofile<-cts5_readProfile(CycleNumber=c,PatternNumber = p,include.inifile=T)

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

dataMerged<-cts5_convert2Ascii(dataprofile)

cts5_save2Ascii(login = login,dataMerged = dataMerged,subdir="./csv/")

}
