% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_decode}
\alias{cts5_decode}
\title{cts5_decode : Use NKE routine to decode .hex data files}
\usage{
cts5_decode(floatname = "ffff", CycleNumber, PatternNumber = 1,
  sensors = c("sbe41", "do", "eco", "ocr", "suna"), subdir = ".",
  OS_windows_progpath = "D:/Data/Provor_USEA/USEA_R/",
  OS_Linux_progpath = "/home/edouard/Documents/APMT-USEA-decod")
}
\arguments{
\item{floatname}{hexa name of the float}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensors}{list of sensor to decode}

\item{subdir}{subdir where to put .csv ASCII files}
}
\value{
No return. This function create ASCII files in the subdir directory
}
\description{
call apmtdecoder.exe routine to decode .hex data files and create NKE .csv ASCII files.
}
\details{
this function must be call in the directory where are .hex files
}
