% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_Energy.R
\name{cts5_energy_decode}
\alias{cts5_energy_decode}
\title{Use NKE routine to estimate energy from technical file}
\usage{
cts5_energy_decode(
  floatname = "",
  techfilename = "",
  CycleNumber,
  PatternNumber = 1,
  subdir = "./eny",
  Nke_ProgPath = ""
)
}
\arguments{
\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{subdir}{subdir where to put .csv ASCII files}

\item{Nke_ProgPath}{path to the nke decoder (APMTEnergy.exe or ...). This path is stored to Sys.getenv("USEAR_Nke_ProgPath").
For Linux, if the Decoder path is in the .bashrc, you can leave this parameters empty.}
}
\value{
No return. This function create ASCII files in the subdir directory
}
\description{
call NKE routine to estimate energy from technical file
}
\details{
this function must be call in the directory where are technicals and ini files
}
\examples{
cts5_energy_decode(CycleNumber=c,PatternNumber = p,subdir="./eny")

cts5_energy_decode(floatname = floatname,CycleNumber=c,PatternNumber=p,subdir="./eny",
Nke_ProgPath="D:/Data/Provor_USEA/USEA_R/")


}
