% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_Energy.R
\name{cts5_energy_plot}
\alias{cts5_energy_plot}
\title{Plot energy estimation as pdf file}
\usage{
cts5_energy_plot(
  login,
  floatname = "",
  subdir = "./ENY",
  metadata = NULL,
  batteryInitialCapacity = 0.9,
  fitLength = 10
)
}
\arguments{
\item{login}{identifiant used at the beginning of the pdf filename}

\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}

\item{subdir}{subdir where energy .csv ASCII files are stored from \code{\link{cts5_energy_process}}}

\item{metadata}{from \code{\link{cts5_readMetaSensor}} to get battery capacity}

\item{batteryInitialCapacity}{initial battery capacity [0-1]}

\item{fitLength}{nombre of profile taken into account to fit the remaining number of profile}
}
\value{
a pdf file in the current directory and a list with various outputs.
}
\description{
open csv file created by \code{\link{cts5_energy_process}} and plot data
}
\details{
this function must be call in the directory where are technical files
}
\examples{
Meta<-cts5_readMetaSensor()

cts5_energy_process(floatname="3ab0",Nke_ProgPath="D:/Data/Provor_USEA/USEA_R/",metadata=Meta)

cts5_energy_plot(login,floatname,metadata=Meta)

}
