% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_Energy.R
\name{cts5_energy_process}
\alias{cts5_energy_process}
\title{Use NKE routine to estimate energy from technical file}
\usage{
cts5_energy_process(
  Energydata = NULL,
  floatname = "",
  subdir = "./eny",
  AutoSaveLoad = F,
  FromLastReset = F,
  CycleNumber = NULL,
  Nke_ProgPath = ""
)
}
\arguments{
\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}

\item{subdir}{subdir where to put .csv ASCII files}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{Nke_ProgPath}{path to the nke decoder (APMTDecrypt.exe or Decoder). This path is stored to Sys.getenv("USEAR_Nke_ProgPath").
For Linux, if the Decoder path is in the .bashrc, you can leave this parameters empty.}

\item{PatternNumber}{numeric : number of the Pattern to decode}
}
\value{
No return. This function create ASCII files in the subdir directory
}
\description{
call NKE routine to estimate energy from technical file
}
\details{
this function must be call in the directory where are technical files
}
\examples{
cts5_energy_process(floatname="3ab0",Nke_ProgPath="D:/Data/Provor_USEA/USEA_R/")

cts5_energy_plot(login,floatname)

}
