% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_NextPatternInfos}
\alias{cts5_NextPatternInfos}
\title{Provides information about the next pattern}
\usage{
cts5_NextPatternInfos(dataprofile)
}
\arguments{
\item{dataprofile}{data and technical files read from \code{\link{cts5_readProfile}}}
}
\value{
list containing the information
\itemize{
 \item{NextSurface_time: }{Estimated date for the next surfacing}
 \item{NextCycle: }{Next Cycle Number}
 \item{NextPattern: }{Next Pattern Number}
 \item{NextDuration: }{Duration (in hours) of the next pattern}
 \item{NextPatternIni: }{Ini section of the Next Pattern}
}
}
\description{
Provides information about the next pattern including the next surface time estimated
}
\details{
must be used where the ini files could be found.
}
\examples{

dataprofile<-cts5_readProfile(CycleNumber=c,PatternNumber = p,include.inifile=T)
NPI<-cts5_NextPatternInfos(dataprofile)

}
