% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_PlotTechnical}
\alias{cts5_PlotTechnical}
\title{Plot Technical informations in one plot}
\usage{
cts5_PlotTechnical(
  tech,
  output = "Plot_technical.pdf",
  floatname = "",
  mfrow = c(4, 3),
  toplot = c("Date", "Pi", "Vbatt(V)", "Volumes", "HydroActions", "EV/Pump", "Depths",
    "Data", "Iridium", "Commands", "Power", "ALARM")
)
}
\arguments{
\item{tech}{technical data read from \code{\link{cts5_readalltech}}}

\item{output}{name of the pdf file}

\item{floatname}{name of the float to add on the plot}

\item{toplot}{list of information to plot}
}
\value{
a data.frame
}
\description{
Plot Technical informations in one plot
}
\examples{
tech<-cts5_readalltech()
cts5_PlotTechnical(tech,floatname=tech$floatname[1])

}
