% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_ProcessData}
\alias{cts5_ProcessData}
\title{Process data to physical values}
\usage{
cts5_ProcessData(metadata, dataprofile, ProcessUncalibrated = F)
}
\arguments{
\item{metadata}{is a list containing technical data including calibration coefficients 
obtained by using \code{\link{cts5_readMetaSensor}}}

\item{dataprofile}{data and technical files read from \code{\link{cts5_readProfile}}}

\item{ProcessUncalibrated}{if True, process physical data with default calibration if needed}
}
\value{
data.frame containing the data
}
\description{
apply processing to raw values in order to compute physical values
}
\details{
calibration coefficients are usually in 00_metadata.xml file.
}
\examples{

login="lovuse001a"

Meta<-cts5_readMetaSensor()

cts5_decode(CycleNumber=c,PatternNumber = p,subdir="./CSV")

dataprofile<-cts5_readProfile(CycleNumber=c,PatternNumber = p,include.inifile=T)

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

PlotCTS5(login=login,dataprofile,PhaseToPlot=c("PRE","DES","PAR","ASC","SUR"),add=FALSE,technical=TRUE,paper = "A4",mfrow=c(3,2))


}
