% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_readcsv}
\alias{cts5_readcsv}
\title{read USEA .csv data file}
\usage{
cts5_readcsv(
  floatname = "ffff",
  CycleNumber,
  PatternNumber = 1,
  sensor = "sbe41",
  dec = "."
)
}
\arguments{
\item{floatname}{hexa name of the float}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensor}{name of the sensor to read from the list CTS5_supported_sensors}

\item{dec}{decimal character in ASCII}
}
\value{
data.frame containing the data
}
\description{
read NKE .csv ASCII files obtained from \code{\link{cts5_decode}} .
}
\details{
this function add an identification for the sensorType from \code{\link{cts5_SensorTypeId}} 
and acquisition phase.


Phase
[DESCENT]->"DES"
[PARK]->"PAR"
[DEEP_PROFILE]->"DEE"
[SHORT_PARK]->"SHP"
[ASCENT]->"ASC"
[SURFACE]->"SUR"
}
