% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_datalogger.R
\name{cts5_readDatalogger}
\alias{cts5_readDatalogger}
\title{cts5_readDatalogger : Concat .csv files to one data.frame}
\usage{
cts5_readDatalogger(filename)
}
\arguments{
\item{filename}{}
}
\value{
a list containing the data with the same format than \code{\link{cts5_readProfile}}
}
\description{
read datalogger .csv ASCII files
}
\examples{

logger<-cts5_readDatalogger("3aa9_datalogger.csv")

Meta<-cts5_readMetaSensor()

logger<-cts5_ProcessData(Meta$SENSORS,logger)

}
