% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_readIni}
\alias{cts5_readIni}
\title{read CTS5 _apmt.ini file}
\usage{
cts5_readIni(inifilename)
}
\arguments{
\item{inifilename}{filename of the ini file}
}
\value{
list containing the information
}
\description{
read CTS5 _apmt.ini file and store the information into a list
}
