% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_readMetaSensor}
\alias{cts5_readMetaSensor}
\title{read CTS5 Metadata file.}
\usage{
cts5_readMetaSensor(
  floatname = "",
  CycleNumber = NULL,
  PatternNumber = NULL,
  filename = NULL
)
}
\arguments{
\item{floatname}{hexa name of the float or .* to select all float name. If "" the floatname of the last file transmitted will be used}

\item{CycleNumber}{numeric : number of the cycle to decode. If NA, 
the last file is decoded}

\item{PatternNumber}{numeric : number of the Pattern to decode. If NA, 
the last file is decoded}

\item{filename}{if not null, used as meta file filename}
}
\value{
list containing the Meta data
}
\description{
read CTS5 xml files with float and sensors meta data
}
\examples{

Meta<-cts5_readMetaSensor(floatname="ffff",CycleNumber=118)

Meta<-cts5_readMetaSensor()

}
