% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_readProfile}
\alias{cts5_readProfile}
\title{cts5_readProfile : Concat .csv files to one data.frame}
\usage{
cts5_readProfile(
  floatname = "",
  CycleNumber,
  PatternNumber = 1,
  sensors = CTS5_supported_sensors,
  dec = ".",
  include.inifile = F,
  csv.subdir = "./CSV"
)
}
\arguments{
\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensors}{list of sensor to decode}

\item{dec}{decimal character in ASCII}

\item{include.inifile}{If True, read the corresponding iniFile and add it to the list.}
}
\value{
a list containing the data
}
\description{
read NKE .csv ASCII files obtained from \code{\link{cts5_decode}} and 
concat them into a list
}
\examples{

login="lovuse001a"

Meta<-cts5_readMetaSensor(floatname=floatname)

cts5_decode(CycleNumber=c,PatternNumber = p,subdir="./CSV")

dataprofile<-cts5_readProfile(CycleNumber=c,PatternNumber = p,include.inifile=T)

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

}
