% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_readtechnical}
\alias{cts5_readtechnical}
\title{read and parse CTS5 technical file.}
\usage{
cts5_readtechnical(
  filename = "",
  floatname = "",
  CycleNumber,
  PatternNumber = 1
)
}
\arguments{
\item{filename}{file to open. If exist, will replace the automatic name provided by Cycle
and Pattern number.}

\item{floatname}{hexa name of the float}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}
}
\value{
list containing the technical data
}
\description{
read and parse CTS5 technical file (_tecnical.txt)
}
