% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_save2Ascii}
\alias{cts5_save2Ascii}
\title{save data to CTS5 csv format}
\usage{
cts5_save2Ascii(login, dataMerged, subdir = "./csv/")
}
\arguments{
\item{login}{login of the float used as prefix}

\item{dataMerged}{Merged data obtained by \code{\link{cts5_convert2Ascii}}}

\item{subdir}{sub directory where to save the data}
}
\description{
save data to CTS5 csv format
}
\examples{

login="lovuse001a"

Meta<-cts5_readMetaSensor()

cts5_decode(CycleNumber=c,PatternNumber = p,subdir="./CSV")

dataprofile<-cts5_readProfile(CycleNumber=c,PatternNumber = p,include.inifile=T)

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

dataMerged<-cts5_convert2Ascii(dataprofile)

cts5_save2Ascii(login = login,dataMerged = dataMerged,subdir="./csv/")

}
