% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_save2RData}
\alias{cts5_save2RData}
\title{cts5_save2RData}
\usage{
cts5_save2RData(dataprofile, login = "", subdir = "")
}
\arguments{
\item{dataprofile}{dataprofile to be saved}

\item{login}{identifiant used at the beginning of the pdf filename}
}
\description{
save dataprofile to RData
}
\examples{
cts5_save2RData(dataprofile,login = login)

cts5_save2RData(dataprofile,login = login,subdir="./csv/")

}
