% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_SaveToTrajFile}
\alias{cts5_SaveToTrajFile}
\title{Sauvegarde des donnees navigation au format csv dans un fichier login_ccc_pp_trajectory.csv}
\usage{
cts5_SaveToTrajFile(login, NavData, split.file = TRUE, show.PumpEV = TRUE)
}
\arguments{
\item{login}{login of th float}

\item{NavData}{NavData read from cts5_system_process}

\item{split.file}{cut files in cycle and pattern}

\item{show.PumpEV}{include Ev and Pump data}
}
\description{
Sauvegarde des donnees navigation au format csv dans un fichier login_ccc_pp_trajectory.csv
}
\examples{
Navdata<-cts5_system_process(floatname = "3ab3")
Navdata<-cts5_system_CutNavdata(Navdata,CycleNumber=12)

Meta<-cts5_readMetaSensor()
login<-Meta$TELECOM["Login"]

cts5_SaveToTrajFile(login,Navdata)

}
