% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_ScanProfilesID}
\alias{cts5_ScanProfilesID}
\title{find sensor list from Meta files}
\usage{
cts5_ScanProfilesID(pattern = ".*sbe41.hex")
}
\arguments{
\item{pattern}{pattern used to list data}
}
\value{
a list of floatname, cycle and pattern
}
\description{
find sensor list from Meta files

cts5_ScanProfilesID scan the directory to list cyle and pattern available
}
\examples{

findfloatname(pattern="[[:alnum:]]{4}_[[:digit:]]{3}_[[:digit:]]{2}.*")



list the data available

cts5_ScanProfilesID(pattern=".*01_technical.txt")

cts5_ScanProfilesID()

}
