% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_SensorTypeId}
\alias{cts5_SensorTypeId}
\title{provide sensor Id used un csv file}
\usage{
cts5_SensorTypeId(pattern = "")
}
\arguments{
\item{pattern}{name of the sensor to look for. if "", provide the full list.}
}
\value{
a vector containing the id of the sensors
}
\description{
provide sensor Id used un csv file \code{\link{cts5_readcsv}} .
}
\details{
SensorType
  0 : CTD
  3 : DO Aanderaa
  9 : ECOpuck
 12 : OCR504
 18 : cROVER
 21 : SUNA
 22 : PHSEABIRD
109 : UVP6 lpm
110 : UVP6 blk
111 : UVP6 TAXO1 (future use)
112 : UVP6 TAXO2 (future use)
113 : Ramses
114 : opus_lgt
115 : opus_blk
}
\examples{
cts5_SensorTypeId("")

cts5_SensorTypeId("uvp6")

cts5_SensorTypeId("uvp6_blk")

}
