% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\docType{data}
\name{CTS5_supported_sensors}
\alias{CTS5_supported_sensors}
\title{list of available sensors}
\format{
An object of class \code{character} of length 14.
}
\usage{
CTS5_supported_sensors
}
\description{
vector of names of available sensors
}
\keyword{datasets}
