% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_CutNavdata}
\alias{cts5_system_CutNavdata}
\title{Cut a NavData file for a given CycleNumber and PatternNumber}
\usage{
cts5_system_CutNavdata(Navdata = NULL, CycleNumber = NULL, PatternNumber = 1)
}
\arguments{
\item{Navdata}{already processed system files}

\item{CycleNumber}{numeric : number of the cycle to extract from}

\item{PatternNumber}{numeric : number of the Pattern to extract from}
}
\value{
Navdata for the given profile
}
\description{
Cut a NavData file for a given CycleNumber and PatternNumber
}
\examples{
Navdata<-cts5_system_process(floatname = "3ab3")
Navdata<-cts5_system_CutNavdata(Navdata,CycleNumber=12)


}
