% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_decode}
\alias{cts5_system_decode}
\title{Use NKE routine to decode .hex system files}
\usage{
cts5_system_decode(
  filename = "",
  floatname = "ffff",
  sysfile_number = NULL,
  CycleNumber = NULL,
  PatternNumber = 1,
  subdir = ".",
  OS_windows_progpath = "D:/Data/Provor_USEA/USEA_R/",
  OS_Linux_progpath = "/home/edouard/Documents/APMT-USEA-decod"
)
}
\arguments{
\item{filename}{filename of the system file}

\item{floatname}{hexa name of the float}

\item{sysfile_number}{numeric : system file number}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{subdir}{not used}

\item{OS_windows_progpath}{path to the nke decoder (APMTDecrypt.exe) for a windows system}

\item{OS_Linux_progpath}{path to the nke decoderfor (APMTDecrypt.exe) a Linux system. !! Attention, required
the installation of wine}
}
\value{
filename of the decoded file
}
\description{
call APMTDecrypt.exe routine to decode .hex system files and create NKE .csv ASCII files.
}
\details{
this function must be call in the directory where are .hex files

This function will decode the system file in the following order of priority :
1- filename if provided, 2- the system file with the sysfile_number if provided, 
3- the system file recorded at the same time than the technical file specified by CycleNumber and PatternNumber
}
\examples{
decrypt_filename<-cts5_system_decode(floatname = "3ab3",sysfile_number=131,
OS_windows_progpath="my path to APMTDecrypt.exe")

Navdata<-cts5_system_parse(decrypt_filename)

cts5_system_plot(Navdata)

}
