% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_merge}
\alias{cts5_system_merge}
\title{Merge Parsed system file}
\usage{
cts5_system_merge(Navdata1, Navdata2)
}
\arguments{
\item{NavData1}{parsed data from \code{\link{cts5_system_parse}}}

\item{NavData2}{parsed data from \code{\link{cts5_system_parse}}}
}
\value{
parsed data
}
\description{
Merge Parsed system file from \code{\link{cts5_system_parse}}
}
\examples{

NavData<-cts5_system_parse("ffff_system_01755_decrypt.log")
NavData2<-cts5_system_parse("ffff_system_01756_decrypt.log")

Navdata<-cts5_system_merge(Navdata,Navdata2)

}
