% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_Plot}
\alias{cts5_system_Plot}
\title{Plot parsed system file}
\usage{
cts5_system_Plot(
  NavData,
  cycle = NULL,
  pattern = 1:10,
  ylim = NULL,
  timeWin = c(0, 1),
  relatif = FALSE,
  DepthPch = 20,
  Depthtype = "b",
  show.EV = TRUE,
  show.pump = TRUE,
  show.title = F,
  show.cycle = T,
  show.user = T,
  legendPos = ""
)
}
\arguments{
\item{NavData}{pased data from \code{\link{cts5_system_parse}}}

\item{cycle}{list of cycle number to plot}

\item{pattern}{pattern to plot}

\item{ylim}{plot ylim}

\item{timeWin}{relative time window.}
}
\value{
list containing the parsed system information
}
\description{
plot trajectory of the float from \code{\link{cts5_system_parse}}
}
\examples{

NavData<-cts5_system_parse("ffff_system_01755_decrypt.log")
NavData<-cts5_system_parse("_system_decrypt.log",userdefine="Grounding detected")

cts5_system_Plot(NavData)
cts5_system_Plot(NavData,cycle=5:10)
cts5_system_Plot(NavData,cycle=5,pattern=1)


}
