% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCTS5_USEA.R
\name{PlotCTS5}
\alias{PlotCTS5}
\title{PlotCTS5 : Plot a profile}
\usage{
PlotCTS5(
  login = "lov",
  dataprofile,
  PhaseToPlot = c("PRE", "DES", "PAR", "ASC", "SUR"),
  add = FALSE,
  technical = TRUE,
  paper = "A4",
  mfrow = c(3, 2)
)
}
\arguments{
\item{login}{identifiant used at the beginning of the pdf filename}

\item{dataprofile}{data and technical files read from \code{\link{cts5_readProfile}}
processing by \code{\link{usea_ProcessData}}}

\item{PhaseToPlot}{list of phase to plot}

\item{add}{if false a new pdf will be created. if true all plots will be generated in the 
current open device}

\item{technical}{if true, technical information will be plotted}

\item{paper}{paper size}

\item{mfrow}{mfrow}
}
\description{
PlotCTS5 : create a pdf file with profile plots
}
\examples{

login="lovuse001a"

floatname="ffff"

Meta<-cts5_readMetaSensor(floatname=floatname)

cts5_decode(floatname=floatname,CycleNumber=c,PatternNumber = p,subdir="./CSV")

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

dataprofile<-cts5_ProcessData(Meta$SENSORS,dataprofile)

PlotCTS5(login=login,dataprofile,PhaseToPlot=c("PRE","DES","PAR","ASC","SUR"),add=FALSE,technical=TRUE,paper = "A4",mfrow=c(3,2))

 

}
