% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{ProjectPosition2Recover}
\alias{ProjectPosition2Recover}
\title{ProjectPosition2Recover : project position based on previous Postions}
\usage{
ProjectPosition2Recover(nextMin = 0, data = data, ind = 1)
}
\arguments{
\item{nextMin}{Number of minutes from now to project the position}

\item{data}{Positions read from \code{\link{ScanPosition2Recover}} or \code{\link{ScanDefault2Recover}}}

\item{ind}{The drift of the float is estimated between the last position and the position [Last-ind]}
}
\value{
projected position
}
\description{
ProjectPosition2Recover project the futur position of the float in nextMin minutes based on previous positions read 
by \code{\link{ScanPosition2Recover}} or \code{\link{ScanDefault2Recover}}
}
\examples{
data<-ScanDefault2Recover()
ProjectPosition2Recover(nextMin=1,data=data) estimate the position in 1 min from now



}
