% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{Recover_PredictPosition}
\alias{Recover_PredictPosition}
\title{Predict position based on previous Positions}
\usage{
Recover_PredictPosition(nextMin = 0, data = data, ind = 1)
}
\arguments{
\item{nextMin}{Number of minutes from now to project the position}

\item{data}{Positions read from \code{\link{Recover_ScanPosition}} or \code{\link{Recover_ScanDefault}}}

\item{ind}{The drift of the float is estimated between the last position and the position [Last-ind]}
}
\value{
projected position
}
\description{
extrapolates the future position of the float in nextMin minutes based on previous positions read 
by \code{\link{Recover_ScanPosition}} or \code{\link{Recover_ScanDefault}}
}
\examples{
data<-Recover_ScanDefault()
Recover_PredictPosition(nextMin=1,data=data) # estimate the position in 1 min from now



}
