% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{Recover_ScanDefault}
\alias{Recover_ScanDefault}
\title{read technical/default files to recover the float.}
\usage{
Recover_ScanDefault(
  pattern = ".*_default_.*.txt",
  Outputfilename = "Positions.txt",
  KMLfile = "Positions.kml"
)
}
\arguments{
\item{pattern}{pattern of the technical/default files}

\item{Outputfilename}{filename of the Positions file}

\item{KMLfile}{name of the KML file generated}
}
\value{
identical as \code{\link{ScanPosition2Recover}}
}
\description{
This function read technical/default files and write the Potions file to be called by \code{\link{ScanPosition2Recover}}
}
\examples{
ScanDefault2Recover()



}
