% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{Recover_ScanPosition}
\alias{Recover_ScanPosition}
\title{read Position file to recover the float.}
\usage{
Recover_ScanPosition(
  filename = "Positions.txt",
  KMLfile = "Positions.kml",
  GPXfile = ""
)
}
\arguments{
\item{filename}{name of the text file which contains the positions.}

\item{KMLfile}{name of the KML file to be generated. No file is 
created if "".}

\item{GPXfile}{name of the GPX file to be generated. No file is 
created if "".}
}
\value{
a data frame which contains the positions in various format and the speed and course of the float.
}
\description{
This function read the position of the float during a recovering. Positions are entered by hand in a text file or 
read from default files by \code{\link{Recover_ScanDefault}}.
}
\details{
The positions must be entered in a text file by copying the [GPS] line of the technical/default file as follow :
UTC=19-11-25 08:02:00 Lat=4315.57717N Long=00658.52514E Clock drift=+0.032 s
UTC=19-11-25 08:16:35 Lat=4315.27231N Long=00657.97657E Clock drift=+0.000 s
}
