% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{ScanPosition2Recover}
\alias{ScanPosition2Recover}
\title{ScanPosition2Recover : read Position file to recover the float.}
\usage{
ScanPosition2Recover(filename = "Positions.txt", KMLfile = "Positions.kml")
}
\arguments{
\item{filename}{name of the text file which contains the positions.}

\item{KMLfile}{name of the KML file generated}
}
\value{
a data frame which contains the positions in various format and the speed and course of the float.
}
\description{
This function read the position of the float during a recovering. Positions are entered by hand in a text file or 
read from default files by \code{\link{ScanDefault2Recover}}.
}
\examples{
The positions must be entered in a text file by copying the [GPS] line of the technical/default file as follow :
UTC=19-11-25 08:02:00 Lat=4315.57717N Long=00658.52514E Clock drift=+0.032 s
UTC=19-11-25 08:16:35 Lat=4315.27231N Long=00657.97657E Clock drift=+0.000 s



}
