% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_AllTech_inTab}
\alias{cts5_AllTech_inTab}
\title{All technical file in a data Frame}
\usage{
cts5_AllTech_inTab(
  pattern = ".*_technical.*.txt",
  CycleNumber = NULL,
  include_tech0 = FALSE
)
}
\arguments{
\item{pattern}{pattern used to select files}

\item{CycleNumber}{vector of cycle number to read. If Null, all technical files are read.}

\item{include_tech0}{not used}
}
\value{
a data.frame
}
\description{
read all _technical.txt files and return results in a data.frame
}
