% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_create_kml}
\alias{cts5_create_kml}
\title{create a KML file from technical files}
\usage{
cts5_create_kml(
  pattern = ".*technical.*.txt",
  output = "PositionAPMT.kml",
  start = 1,
  CycleToProcess = NULL,
  path = ".",
  id = "cycle"
)
}
\arguments{
\item{pattern}{pattern used to select files}

\item{output}{name of the KML file}

\item{start}{First Cycle number to process.}

\item{CycleToProcess}{vector of float cycle to include in the KML file. If Null, all cycle will be included}
}
\value{
a KML file
}
\description{
read  _technical.txt files and create a KML file to be read with Google Earth
}
\details{
CTS5_create_kml()
}
