% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_create_kml}
\alias{cts5_create_kml}
\title{create a KML or GPX file from technical files}
\usage{
cts5_create_kml(
  pattern = ".*autotest.*.txt|.*technical.txt|.*default.*.txt",
  filenamelist = NULL,
  output = "PositionAPMT.kml",
  start = 1,
  CycleToProcess = NULL,
  path = ".",
  id = "cycle",
  color = "ff00ffff",
  outputtype = "kml",
  namevector = NULL
)
}
\arguments{
\item{pattern}{pattern used to select files}

\item{filenamelist}{list of files to open}

\item{output}{name of the KML or GPX file}

\item{start}{First Cycle number to process.}

\item{CycleToProcess}{vector of float cycle to include in the KML file. If Null, all cycle will be included}

\item{id}{id to identify points in the file, either : "date", 
"cycle" or  "relativecycle"}

\item{color}{: Color of points in KML file
Red = ff0000ff,
Yellow = ff00ffff,
Blue = ffff0000,
Green = ff00ff00,
Purple = ff800080,
Orange = ff0080ff,
Brown = ff336699,
Pink = ffff00ff}

\item{namevector}{: vector of names to be used as point's name.}

\item{type}{"kml" or "gpx" : choose file type}
}
\value{
a KML or GPX file
}
\description{
read  technical, autotest or default files and create a KML or GPX
file to be read with Google Earth, OpenCPN ...
}
\examples{
 

# Automatic
cts5_create_kml()

# with filenamelist
tech<-cts5_readalltech()
cts5_create_kml(filenamelist=tech$filename)
 

}
