% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_decode}
\alias{cts5_decode}
\title{Use NKE routine to decode .hex data files}
\usage{
cts5_decode(
  floatname = "",
  CycleNumber,
  PatternNumber = 1,
  sensors = CTS5_supported_sensors,
  subdir = ".",
  Nke_ProgPath = ""
)
}
\arguments{
\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensors}{list of sensor to decode. From the list CTS5_supported_sensors}

\item{subdir}{subdir where to put .csv ASCII files}

\item{Nke_ProgPath}{path to the nke decoder (APMTDecrypt.exe or Decoder). This path is stored to Sys.getenv("USEAR_Nke_ProgPath").
For Linux, if the Decoder path is in the .bashrc, you can leave this parameters empty.}
}
\value{
No return. This function create ASCII files in the subdir directory
}
\description{
call apmtdecoder.exe routine to decode .hex data files and create NKE .csv ASCII files.
}
\details{
this function must be call in the directory where are .hex files
}
