% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_Energy.R
\name{cts5_energy_plot}
\alias{cts5_energy_plot}
\title{Plot energy estimation as pdf file}
\usage{
cts5_energy_plot(login, floatname, IsJumbo = F, subdir = "./eny")
}
\arguments{
\item{login}{identifiant used at the beginning of the pdf filename}

\item{floatname}{hexa name of the float. If "", the floatname will automatically found.}
}
\value{
a pdf file in the current directory
}
\description{
open csv file created by \code{\link{cts5_energy_process}}
}
\details{
this function must be call in the directory where are technical files
}
\examples{
cts5_energy_process(floatname="3ab0",Nke_ProgPath="D:/Data/Provor_USEA/USEA_R/")

cts5_energy_plot(login,floatname)

}
