% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_ProcessData}
\alias{cts5_ProcessData}
\title{Process data to physical values}
\usage{
cts5_ProcessData(
  metadata,
  dataMerged,
  sensor = c("eco", "ocr", "do", "suna", "sbeph")
)
}
\arguments{
\item{metadata}{is a list containing technical data including calibration coefficients 
obtained by using \code{\link{cts5_readMetaSensor}}}

\item{dataMerged}{is the merged data.frame obtained by \code{\link{cts5_concatProfile}}}

\item{sensors}{list of sensor to process}
}
\value{
data.frame containing the data
}
\description{
apply processing to raw values in order to compute physical values
}
\details{
calibration coefficients are usually in 00_technical.txt file.
}
\examples{

Meta<-cts5_readMetaSensor()

tech<-cts5_readtechnical(floatname=floatname,CycleNumber=c,PatternNumber = p)

cts5_decode(floatname=floatname,CycleNumber=c,PatternNumber = p,subdir="./CSV")

setwd("./CSV")

dataMerged<-cts5_concatProfile(floatname=floatname,CycleNumber=c,PatternNumber = p)

dataMerged<-cts5_ProcessData(Meta$SENSORS,dataMerged)

PlotCTS5(login=login,dataMerged,PhaseToPlot=c("PRE","DES","PAR","ASC","SUR"),add=FALSE,technical=TRUE,paper = "A4",mfrow=c(3,2))

SaveToCTS5(login = login,dataMerged = dataMerged,GPS = tech$GPS)

}
