% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_ProcessData}
\alias{cts5_ProcessData}
\title{Process data to physical values}
\usage{
cts5_ProcessData(
  metadata,
  dataMerged,
  sensor = c("eco", "ocr", "do", "suna", "sbeph")
)
}
\arguments{
\item{metadata}{is a list containing technical data including calibration coefficients 
obtained by using \code{\link{cts5_readMetaSensor}}}

\item{dataMerged}{is the merged data.frame obtained by \code{\link{cts5_concatProfile}}}

\item{sensors}{list of sensor to process}
}
\value{
data.frame containing the data
}
\description{
apply processing to raw values in order to compute physical values
}
\details{
calibration coefficients are usually in 00_technical.txt file.
}
\examples{
dataMerged<-usea_concatProfile(floatname="ffff",CycleNumber=275)

}
