% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_readalltech}
\alias{cts5_readalltech}
\title{All technical file in a data Frame}
\usage{
cts5_readalltech(
  pattern = ".*_technical.*.txt",
  filenames = NULL,
  CycleNumber = NULL,
  include_tech0 = FALSE,
  FromLastReset = FALSE
)
}
\arguments{
\item{pattern}{pattern used to select files}

\item{filenames}{vector of files to read. exclude all other options}

\item{CycleNumber}{vector of cycle number to read. If Null, all technical files are read.}

\item{include_tech0}{If True, include 00_technical files}

\item{FromLastReset}{IF True, start from the last 00_technical files}
}
\value{
a data.frame
}
\description{
read all _technical.txt files and return results in a data.frame
}
\examples{

## Scan Alarms
tech<-cts5_readalltech()
cbind(tech$Cycle_Number,tech$Pattern_Number,tech[,grep("ALARM",colnames(tech))])


}
