% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{cts5_readcsv}
\alias{cts5_readcsv}
\title{read USEA .csv data file}
\usage{
cts5_readcsv(
  floatname = "ffff",
  CycleNumber,
  PatternNumber = 1,
  sensor = "sbe41",
  dec = "."
)
}
\arguments{
\item{floatname}{hexa name of the float}

\item{CycleNumber}{numeric : number of the cycle to decode}

\item{PatternNumber}{numeric : number of the Pattern to decode}

\item{sensor}{name of the sensor to read from the list c("sbe41","do","eco","ocr")}

\item{dec}{decimal character in ASCII}
}
\value{
data.frame containing the data
}
\description{
read NKE .csv ASCII files obtained from \code{\link{cts5_decode}} .
}
\details{
this function add an identification for the sensorType and acquisition phase.

 SensorType
  0 : CTD
  3 : DO Aanderaa
  9 : ECOpuck
 12 : OCR504
 21 : SUNA
 22 : PHSEABIRD
101 : OCR507 1
102 : OCR507 2
103 : Tilt
104 : PSA916
105 : RTSys
106 : Contros
107 : OptTak
108 : Trios_Nitrate
109 : Octopus lpm
110 : Octopus blk

Phase
[DESCENT]->"DES"
[PARK]->"PAR"
[DEEP_PROFILE]->"DEE"
[SHORT_PARK]->"SHP"
[ASCENT]->"ASC"
[SURFACE]->"SUR"
}
