% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA_technical.R
\name{cts5_readMetaSensor}
\alias{cts5_readMetaSensor}
\title{cts5_readMetaSensor : read APMT technical file.}
\usage{
cts5_readMetaSensor(floatname = ".*", CycleNumber = NA, PatternNumber = 0)
}
\arguments{
\item{floatname}{hexa name of the float or .* to select all float name}

\item{CycleNumber}{numeric : number of the cycle to decode. If NA, 
the last file is decoded}

\item{PatternNumber}{numeric : number of the Pattern to decode}
}
\value{
list containing the Meta data
}
\description{
read APMT xml files with sensor meta data
}
\examples{

Meta<-cts5_readMetaSensor(floatname="ffff",CycleNumber=118)

Meta<-cts5_readMetaSensor()

}
