% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_parse}
\alias{cts5_system_parse}
\title{Parse CTS5 system file.}
\usage{
cts5_system_parse(filename = "_system_decrypt.log", userdefine = "")
}
\arguments{
\item{filename}{filename of the decrypted system file (*.log)}

\item{userdefine}{Key word to look for in the system file}
}
\value{
list containing the parsed system information
}
\description{
parse the cts5 system file after decrypting
}
