% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_System.R
\name{cts5_system_Plot}
\alias{cts5_system_Plot}
\title{Plot parsed system file}
\usage{
cts5_system_Plot(
  NavData,
  cycle = NULL,
  pattern = 1:10,
  xlim = NULL,
  ylim = NULL,
  timeWin = c(0, 1),
  relatif = FALSE,
  timeRange = NULL,
  DepthPch = 1,
  PumpPch = 2,
  EVPch = 3,
  Depthtype = "b",
  show.EV = TRUE,
  show.pump = TRUE,
  show.title = F,
  show.cycle = T,
  show.user = T,
  legendPos = "",
  add = F
)
}
\arguments{
\item{NavData}{pased data from \code{\link{cts5_system_parse}}}

\item{cycle}{list of cycle number to plot}

\item{pattern}{pattern to plot}

\item{xlim}{for plot function}

\item{ylim}{plot ylim}

\item{timeWin}{relative time window (if xlim=NULL) expressed as a vector of 2 floats. timeWin=c(0,1) is the full window.}

\item{relatif}{If True, time in min is relative to the first date}

\item{timeRange}{vector of two dates that delimit the plotted data.}
}
\value{
list containing the parsed system information

NavData<-cts5_system_parse("ffff_system_01755_decrypt.log")
NavData<-cts5_system_parse("_system_decrypt.log",userdefine="Grounding detected")

cts5_system_Plot(NavData)
cts5_system_Plot(NavData,cycle=5:10)
cts5_system_Plot(NavData,cycle=5,pattern=1)

## To compare profiles
cts5_system_Plot(Navdata,cycle = 46,relatif = T,xlim=c(0,1500),show.cycle = F)
cts5_system_Plot(Navdata,cycle = 49,relatif = T,show.cycle = F,add=T)

## To plot an Ice braking
Brakedate<-NavData$BrakeRecord$date[1]
cts5_system_Plot(NavData,timeRange=c(Brakedate-1800,Brakedate+1800))
}
\description{
plot trajectory of the float from \code{\link{cts5_system_parse}}
}
