% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5ScanPositionRecover.R
\name{Recover_ScanDefault}
\alias{Recover_ScanDefault}
\title{read technical/default files to recover the float.}
\usage{
Recover_ScanDefault(
  pattern = ".*_default_.*.txt",
  Outputfilename = "Positions.txt",
  KMLfile = "Positions.kml",
  GPXfile = ""
)
}
\arguments{
\item{pattern}{pattern of the technical/default files}

\item{Outputfilename}{filename of the Positions file}

\item{KMLfile}{name of the KML file to be generated. No file is 
created if "".}

\item{GPXfile}{name of the GPX file to be generated. No file is 
created if "".}
}
\value{
identical as \code{\link{Recover_ScanPosition}}
}
\description{
This function read technical/default files and write the Postions file to be called by \code{\link{Recover_ScanPosition}}
}
\examples{
Recover_ScanDefault()
Recover_ScanDefault(pattern=".*technical.*.txt")
Recover_ScanDefault(pattern=".*(technical|default).*.txt")

}
