% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_UVP6.R
\name{Save_UVP62EcoTaxa}
\alias{Save_UVP62EcoTaxa}
\title{save UVP6 data to EcoPart csv format}
\usage{
Save_UVP62EcoTaxa(login, dataprofile, UVP6_HW_CONF = NULL, subdir = "")
}
\arguments{
\item{login}{login of the float used as prefix}

\item{dataprofile}{Merged data obtained by \code{\link{cts5_readProfile}}}

\item{subdir}{sub directory where to save the data}
}
\description{
save UVP6 data to EcoPart csv format
}
\examples{

Save_UVP62EcoTaxa(login,dataprofile,Meta$SENSORS$SENSOR_UVP6$HW_CONF,subdir="./uvp6/")



}
