% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS5_USEA.R
\name{SaveToCTS5}
\alias{SaveToCTS5}
\title{save data to CTS5 csv format}
\usage{
SaveToCTS5(login, dataMerged, subdir = ".", GPS = NULL)
}
\arguments{
\item{login}{login of the float used as prefix}

\item{dataMerged}{Merged data obtained by \code{\link{cts5_concatProfile}}}

\item{subdir}{sub directory where to save the data}

\item{GPS}{GPS data to be included in the csv. obtained by \code{\link{cts5_readtechnical}}}
}
\description{
save data to CTS5 csv format
}
